Changed variable name
