+++ Harvest.Web/ClientApp/src/Projects/ProjectDetailContainer.test.tsx
@@ -106,7 +106,7 @@ describe("Project Detail Container", () => {
     expect(fieldTitle).toContain("Field Request #3");
   });
 
-  it("Display correct number of attachments", async () => {    
+  it("Display correct number of attachments", async () => {
     await act(async () => {
       global.fetch = jest
         .fn()
@@ -128,9 +128,65 @@ describe("Project Detail Container", () => {
       );
     });
 
-    const attachmentList = container.querySelector(".no-list-style");
-    const attachemntsLength = attachmentList?.getElementsByTagName("li");
+    const attachedFilesList = container.querySelector(".no-list-style");
+    const attachemntsLength = attachedFilesList?.getElementsByTagName("li");
 
     expect(attachemntsLength?.length).toBe(2);
+  });
+
+  it("Display correct number of recent tickets", async () => {
+    await act(async () => {
+      global.fetch = jest
+        .fn()
+        .mockImplementationOnce(() => Promise.resolve(projectResponse))
+        .mockImplementationOnce(() => Promise.resolve(unbilledResponse))
+        .mockImplementationOnce(() => Promise.resolve(ticketResponses))
+        .mockImplementationOnce(() => Promise.resolve(invoiceResponse))
+        .mockImplementationOnce(() => Promise.resolve(fileResponse));
+
+      render(
+        <AppContext.Provider value={(global as any).Harvest}>
+          <MemoryRouter initialEntries={["/project/details/3"]}>
+            <Route path="/project/details/:projectId">
+              <ProjectDetailContainer />
+            </Route>
+          </MemoryRouter>
+        </AppContext.Provider>,
+        container
+      );
+    });
+
+    const ticketTable = document.querySelectorAll("tbody")[1];
+    const rows = ticketTable?.querySelectorAll(".rt-tr-group");
+
+    expect(rows?.length).toBe(3);
+  });
+
+  it("Display correct number of recent invoices", async () => {
+    await act(async () => {
+      global.fetch = jest
+        .fn()
+        .mockImplementationOnce(() => Promise.resolve(projectResponse))
+        .mockImplementationOnce(() => Promise.resolve(unbilledResponse))
+        .mockImplementationOnce(() => Promise.resolve(ticketResponses))
+        .mockImplementationOnce(() => Promise.resolve(invoiceResponse))
+        .mockImplementationOnce(() => Promise.resolve(fileResponse));
+
+      render(
+        <AppContext.Provider value={(global as any).Harvest}>
+          <MemoryRouter initialEntries={["/project/details/3"]}>
+            <Route path="/project/details/:projectId">
+              <ProjectDetailContainer />
+            </Route>
+          </MemoryRouter>
+        </AppContext.Provider>,
+        container
+      );
+    });
+
+    const invoiceTable = document.querySelectorAll("tbody")[0];
+    const rows = invoiceTable?.querySelectorAll(".rt-tr-group");
+
+    expect(rows?.length).toBe(3);
   });
 });
